Seperate bin from module and remove lodash dependency with modern JavaScript
