+++ lib/utils/fixedCharCodeAt.js
@@ -1,3 +1,5 @@
+'use strict';
+
 /**
  * Char to Int, with fix for big numbers
  */
