+++ lib/utils/fixedFromCharCode.js
@@ -1,3 +1,5 @@
+'use strict';
+
 /**
  * Int to char, with fix for big numbers
  * @see https://developer.mozilla.org/en/JavaScript/Reference/Global_Objects/String/fromCharCode
