+++ lib/utils/loadFontelloData.js
@@ -1,3 +1,5 @@
+'use strict';
+
 const SvgPath = require('svgpath');
 
 const fixedFromCharCode = require('./fixedFromCharCode');
