+++ lib/utils/loadSvgData.js
@@ -1,3 +1,5 @@
+'use strict';
+
 const XMLDOMParser = require('xmldom').DOMParser;
 const SvgPath = require('svgpath');
